Lazily built indices lookup
As part of this change move the alias validation the occurs
as part of building the indices lookup to the places where
aliases are created (MetadataCreateIndexService / MetadataIndexAliasesService).

Relates to #77466
